# Conflicts:
#	app/assets/javascripts/discourse/app/lib/transformer/registry.js

+++ app/services/admin_notices/dismiss.rb
@@ -3,10 +3,13 @@
 class AdminNotices::Dismiss
   include Service::Base
 
-  model :admin_notice
-
   policy :invalid_access
+  params do
+    attribute :id, :integer
 
+    validates :id, presence: true
+  end
+  model :admin_notice, optional: true
   transaction do
     step :destroy
     step :reset_problem_check
@@ -14,19 +17,23 @@ class AdminNotices::Dismiss
 
   private
 
-  def fetch_admin_notice(id:)
-    AdminNotice.find_by(id: id)
-  end
-
   def invalid_access(guardian:)
     guardian.is_admin?
   end
 
+  def fetch_admin_notice(params:)
+    AdminNotice.find_by(id: params.id)
+  end
+
   def destroy(admin_notice:)
+    return if admin_notice.blank?
+
     admin_notice.destroy!
   end
 
   def reset_problem_check(admin_notice:)
+    return if admin_notice.blank?
+
     ProblemCheckTracker.find_by(identifier: admin_notice.identifier)&.reset
   end
 end
